upd: 敵を生成するクラスを作成
