+++ src/Game.js
@@ -36,35 +36,6 @@ export class Game {
       new PlayerShip(center.x, center.y, 50, 50, 'blue', 100, 0, 15)
     );
 
-    // fot test: add enemy
-    const createEnemy = () => {
-      const { width: canvasWidth, height: canvasHeight } = this.getCanvasSize();
-      const direction = Math.random() > 0.5 ? 'left' : 'right';
-      const y = Math.random() * canvasHeight;
-
-      if (direction === 'left') {
-        const targetX = canvasWidth / 5;
-        const enemyHead = new LeftEnemyHead(y, targetX, 1);
-        const enemyBody = new EnemyPart(50, 50, "red", 100, 0, 15, 0, 0);
-        const enemyCloseBtn = new EnemyPart(10, 10, "black", 100, 0, 15, 45, -5);
-        enemyHead.registerChild(enemyBody);
-        enemyHead.registerChild(enemyCloseBtn);
-        this.gameState.registerObject(enemyHead);
-        this.gameState.registerObject(enemyBody);
-        this.gameState.registerObject(enemyCloseBtn);
-      } else {
-        const targetX = canvasWidth * 4 / 5;
-        const enemyHead = new RightEnemyHead(y, targetX, 1);
-        const enemyBody = new EnemyPart(50, 50, "red", 100, 0, 15, 0, 0);
-        const enemyCloseBtn = new EnemyPart(10, 10, "black", 100, 0, 15, 45, -5);
-        enemyHead.registerChild(enemyBody);
-        enemyHead.registerChild(enemyCloseBtn);
-        this.gameState.registerObject(enemyHead);
-        this.gameState.registerObject(enemyBody);
-        this.gameState.registerObject(enemyCloseBtn);
-      }
-    };
-
     for (let i = 0; i < 3; i++) {
       createEnemy();
     }
